Extract classes out to separate files, and put them in appropriate packages
