+++ posts/src/main/java/com/neaniesoft/vermilion/posts/data/PostRecordConverters.kt
@@ -28,9 +28,9 @@ import com.neaniesoft.vermilion.posts.domain.entities.TextPostSummary
 import com.neaniesoft.vermilion.posts.domain.entities.ThumbnailSummary
 import com.neaniesoft.vermilion.posts.domain.entities.UriImage
 import com.neaniesoft.vermilion.posts.domain.entities.VideoPostSummary
-import com.neaniesoft.vermilion.ui.videos.VideoDescriptor
-import com.neaniesoft.vermilion.ui.videos.VideoHeight
-import com.neaniesoft.vermilion.ui.videos.VideoWidth
+import com.neaniesoft.vermilion.ui.videos.direct.VideoDescriptor
+import com.neaniesoft.vermilion.ui.videos.direct.VideoHeight
+import com.neaniesoft.vermilion.ui.videos.direct.VideoWidth
 import org.commonmark.node.Document
 import org.commonmark.parser.Parser
 import java.time.Clock
