+++ posts/src/main/java/com/neaniesoft/vermilion/posts/domain/routing/CustomVideoMatchResult.kt
@@ -1,0 +1,6 @@
+package com.neaniesoft.vermilion.posts.domain.routing
+
+sealed class CustomVideoMatchResult {
+    object NoMatch : CustomVideoMatchResult()
+    data class RouteMatch(val route: String) : CustomVideoMatchResult()
+}
