+++ posts/src/main/java/com/neaniesoft/vermilion/posts/domain/routing/CustomVideoRouteMatcher.kt
@@ -1,0 +1,7 @@
+package com.neaniesoft.vermilion.posts.domain.routing
+
+import android.net.Uri
+
+interface CustomVideoRouteMatcher {
+    fun match(linkUri: Uri): CustomVideoMatchResult
+}
