+++ ui/src/main/java/com/neaniesoft/vermilion/ui/videos/direct/VideoDialog.kt
@@ -6,6 +6,7 @@ import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.neaniesoft.vermilion.ui.images.rememberZoomableState
+import com.neaniesoft.vermilion.ui.videos.direct.VideoDescriptor
 
 @ExperimentalMaterialApi
 @Composable
