+++ ui/src/main/java/com/neaniesoft/vermilion/ui/videos/external/redgifs/GifResponse.kt
@@ -1,0 +1,7 @@
+package com.neaniesoft.vermilion.ui.videos.external.redgifs
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class GifResponse(
+    @JsonProperty("gif") val gif: GifInfo
+)
