+++ ui/src/main/java/com/neaniesoft/vermilion/ui/videos/external/resolver/VideoUriResolver.kt
@@ -1,0 +1,9 @@
+package com.neaniesoft.vermilion.ui.videos.external.resolver
+
+import android.net.Uri
+import com.github.michaelbull.result.Result
+
+interface VideoUriResolver {
+    suspend fun resolve(uri: Uri): Result<Uri, VideoResolverError>
+    fun handles(uri: Uri): Boolean
+}
